Create 03_entity.py
Now, this simple code, in larger scales, can play a critical role in how machines read texts, or, in other words, how a random text can be made more machine-readable. Imagine, you would give a random sentence to an AI chatbot without asking for anything specific about it. In this case, a NLP-trained machine will try to put semantic tags on the recognizable tokens (entity tags like proper noun, place, dates etc.) to identify what you might have in mind. For example, in this code, the machine realizes the Barack Obama to be a former U.S. president and that probably the user is looking for some biographical information about him.