Create 04_sensitive_content_clarification_nlp.py
For writing this, I got a lot of help from ChatGPT 4 model. It is a proposal regarding using sensitive terms in a text such as a comment made by someone on Instagram. This block of code can, for example, recognize the entity 'isis' regardless of its case and append a certain comment, as an adjective clause, so that it is made clear to anyone that the mentioned organization is a criminal one. This is a rudimentary concept, but it allows for less censorship on content and more freedom of moderated expression.